requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.21.0
pandas==2.2.2
tqdm==4.66.4
pytest==8.2.0
numpy==1.26.4
llama-index==0.10.20
llama-index-embeddings-huggingface==0.1.3
llama-index-vector-stores-chroma==0.1.3
onnxruntime>=1.17.0,<2.0.0
sentence-transformers==2.7.0
torch==2.2.2
transformers==4.39.3
huggingface_hub>=0.20.0
chromadb==0.4.24
Flask==3.1.1
supabase>=2.0.0
python-dotenv>=1.0.0
wikipedia>=1.4.0
flask-cors>=4.0.0
